+++ Dyalect/Runtime/Types/DyIterator.cs
@@ -323,6 +323,9 @@ namespace Dyalect.Runtime.Types
 
         protected override DyFunction GetStaticMember(string name, ExecutionContext ctx)
         {
+            if (name == "Iterator")
+                return DyForeignFunction.Static(name, Concat, 0, new Par("values", true));
+
             if (name == "concat")
                 return DyForeignFunction.Static(name, Concat, 0, new Par("values", true));