+++ Dyalect/Runtime/Types/DyNil.cs
@@ -39,5 +39,13 @@
         protected override DyObject NotOp(DyObject arg, ExecutionContext ctx) => DyBool.True;
 
         protected override DyObject ToStringOp(DyObject arg, ExecutionContext ctx) => new DyString("nil");
+
+        protected override DyFunction GetStaticMember(string name, ExecutionContext ctx)
+        {
+            if (name == "Nil")
+                return DyForeignFunction.Static(name, _ => DyNil.Instance);
+
+            return null;
+        }
     }
 }